feat: statistics 쿼리 레파지토리를 인터페이스 -> 클래스 형식으로 변경
